rl_query_optimizer/requirements.txt
psycopg2-binary>=2.9
torch>=1.13.0
gymnasium>=0.28.1
networkx>=3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
sqlparse>=0.4.4
